+++ src/Sharpen.Engine/SharpenSuggestions/CSharp80/SwitchExpressions/Suggestions/ReplaceSwitchStatementContainingOnlyReturnsWithSwitchExpression.cs
@@ -1,0 +1,15 @@
+﻿namespace Sharpen.Engine.SharpenSuggestions.CSharp80.SwitchExpressions.Suggestions
+{
+    internal sealed class ReplaceSwitchStatementContainingOnlyReturnsWithSwitchExpression : ISharpenSuggestion
+    {
+        private ReplaceSwitchStatementContainingOnlyReturnsWithSwitchExpression() { }
+
+        public string MinimumLanguageVersion { get; } = CSharpLanguageVersions.CSharp80;
+
+        public ICSharpFeature LanguageFeature { get; } = CSharpFeatures.SwitchExpressions.Instance;
+
+        public string FriendlyName { get; } = "Replace switch statement containing only returns with switch expression";
+
+        public static readonly ReplaceSwitchStatementContainingOnlyReturnsWithSwitchExpression Instance = new ReplaceSwitchStatementContainingOnlyReturnsWithSwitchExpression();
+    }
+}